refactor[#523] : 주소 데이터를 요청하는 커스텀 auth -> map으로 이동
