+++ src/entities/auth/api/index.ts
@@ -1,2 +1,1 @@
 export * from "./getMyInfo";
-export * from "./getRegion";
