+++ src/entities/auth/constants/index.ts
@@ -8,11 +8,4 @@ export const SOCIAL_TYPE = {
   GOOGLE: "구글",
 };
 
-export const REGION_END_POINT = {
-  CURRENT_POSITION: ({ lat, lng }: { lat: number; lng: number }) =>
-    `${API_BASE_URL}/addresses/search-by-location?lat=${lat}&lng=${lng}`,
-  REGION_LIST: (keyword: string) =>
-    `${API_BASE_URL}/addresses?keyword=${keyword}`,
-};
-
 export * from "./queryKey";
