+++ src/entities/map/api/queryKey.ts
@@ -1,0 +1,7 @@
+export const regionQueryKey = {
+  regionAll: () => ["regions"] as const,
+  regionKeyword: (keyword: string) =>
+    [...regionQueryKey.regionAll(), { keyword }] as const,
+  regionLatLng: (latLng: { lat: number; lng: number }) =>
+    [...regionQueryKey.regionAll(), { latLng }] as const,
+} as const;
