+++ src/entities/map/api/useRegionByLatLng.ts
@@ -1,0 +1,31 @@
+import { useQuery } from "@tanstack/react-query";
+import { apiClient } from "@/shared/lib";
+import { REGION_END_POINT } from "../constants";
+import { LatLng } from "../types/client";
+import type { Region } from "../types/server";
+import { regionQueryKey } from "./queryKey";
+
+type GetRegionByLatLngResponse = Region[];
+
+interface GetRegionByLatLngRequest extends NonNullableObject<LatLng> {
+  enabled: boolean;
+}
+
+export const useGetRegionByLatLng = ({
+  lat,
+  lng,
+  enabled,
+}: GetRegionByLatLngRequest) => {
+  return useQuery({
+    queryKey: regionQueryKey.regionLatLng({ lat, lng }),
+    queryFn: () =>
+      apiClient.get<GetRegionByLatLngResponse>(
+        REGION_END_POINT.CURRENT_POSITION({
+          lat,
+          lng,
+        }),
+      ),
+    enabled,
+    staleTime: 1000 * 60 * 5,
+  });
+};
