+++ src/entities/map/constants/index.ts
@@ -1,1 +1,2 @@
 export * from "./tiling";
+export * from "./endPoint";
