+++ src/entities/map/types/server.ts
@@ -1,0 +1,7 @@
+export interface Region {
+  id: number;
+  province: string;
+  cityCounty: string;
+  district: string;
+  subDistrict: string;
+}
